Bug fixes and adjustments
Change position of where 2UP player text was being drawn and player lasers (looked off-centre previously). Change position of text which drew the PAUSED screen.
Bug fixes of sound for thrusting and the if statment which checks if the game is over.
Change asset which is used for the exhaust for player 2. 
Add hitTest for player 2 - hit detection not working properly so this will need to be fixed.